interactive_guess_characters: Give acoustic feedback
Help the user by giving affirmative feedback on correct answers
and a try-again feedback if a wrong answer was given.

The implementation is a hack. Therefore I increased the duration
to wait while the character to be played in CharacterSoundPlayer,
so there is enough room between "richtig" and a potentially
following character.